// src/app/models/form-simulasi.model.ts
export class FormSimulasi {
  id?: string;
  nomorRekening?: string;
  nomorPin?: string;
  tanggalPengajuan?: string;
  perihalUpgrade?: string;
  tanggalRealisasi?: string;
  tahunRealisasi?: string;
  nama?: string;
  produk?: string;
  otr?: number;
  merkKendaraan?: string;
  kategoriKendaraan?: string;
  tahunKendaraan?: string;
  jenisPenggunaan?: string;
  maskapaiAsuransi?: string;
  kondisi?: string;
  noPolis?: string;
  noPolisi?: string;
  coverMundur?: Boolean;
  survey?: Boolean;
  RSCC?: number;
  bencanaAlam?: number;
  TJH?: number;
  aksesoris?: number;
  upgradeAllRisk?: Boolean;
  periodeUpgradeDari?: number;
  periodeUpgradeSampai?: number;
  biayaLoading?: boolean;
  alamatPengiriman?: string;
  remarks?: string;
  biayaAdmin: number = 14000;
  totalPremiPertanggungan: number;
  totalPremiPerluasan: number ;
  totalBiaya: number = 14000;
  statusUpgradeId: number;
  va: string;
  dateExpired: string;
}